MIS581 Project SAS Studio  Code
The code from lines 1 through 52 is used to convert the file types pulled from the NHANES website, merge the datasets by participant identifier, and prepare the dataset by removing missing values. The provided dataset in the README is the final product of the code execution for lines 1 through 52. All subsequent code was used to analyze the data provided in the README.